Move is_child_of to call site
